Extract validateConfig() to separate file
